move abt_supervisor expression out of core
